+++ ifinitySDK.framework/Versions/A/Headers/IFMVenue.h
@@ -1,9 +1,9 @@
 //
 //  IFMVenue.h
-//  ifinitySDK
+//  Pods
 //
-//  Created by Wojciech Chojnacki on 14.04.2015.
-//  Copyright (c) 2015 GetIfinity. All rights reserved.
+//  Created by Mario on 06/07/15.
+//
 //
 
 #import <Foundation/Foundation.h>
@@ -15,11 +15,13 @@
 
 @property (nonatomic, retain) NSNumber * center_lat;
 @property (nonatomic, retain) NSNumber * center_lng;
+@property (nonatomic, retain) NSString * ibeaconUUID;
 @property (nonatomic, retain) NSString * name;
 @property (nonatomic, retain) NSNumber * publicVenue;
 @property (nonatomic, retain) NSNumber * remote_id;
 @property (nonatomic, retain) NSString * type;
 @property (nonatomic, retain) NSNumber * update_time;
+@property (nonatomic, retain) NSNumber * outdoor;
 @property (nonatomic, retain) NSSet *beacons;
 @property (nonatomic, retain) IFMContent *content;
 @property (nonatomic, retain) NSSet *floorplans;